frontend authentication complete

+++ src/components/WorkoutDetails.jsx
@@ -1,16 +1,25 @@
 /* eslint-disable react/prop-types */
 import { useContext } from "react";
 import { WorkoutsContext } from "../context/WorkoutContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 // date fns
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 export default function WorkoutDetails({workout}) {
   const { dispatch } = useContext(WorkoutsContext);
+  const {user} = useAuthContext();
 
   const handleClick = async () => {
+    if(!user) {
+      return
+    }
+
     const response = await fetch('http://localhost:4000/api/workouts/' + workout._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     })
     
     const json = await response.json();
